Ingest to Weaviate Images/requirements.txt
quixstreams==2.5.1
python-dotenv
weaviate-client==4.6.4